Bundle. Start message for chat.
